Allow errors to propagate to other steps
(WHAT)
It is required to reuse the error messages in other steps in the
workflow. This will allow showing the error messages as a sticky
comments in the PR

(HOW)
You can use jobs.<job_id>.outputs to create a map of outputs for
a job. Job outputs are available to all downstream jobs that
depend on this job.
See (https://docs.github.com/en/actions/using-jobs/defining-outputs-for-jobs)